feat: working in requested changes
