+++ src/app.module.ts
@@ -25,7 +25,6 @@ import { AppSumoModule } from './modules/app-sumo/app-sumo.module';
 import { PlanModule } from './modules/plan/plan.module';
 import { WorkspacesModule } from './modules/workspaces/workspaces.module';
 import { GatewayModule } from './modules/gateway/gateway.module';
-import { TwoFactorAuthModule } from './modules/tfa/two-factor-auth.module';
 
 @Module({
   imports: [
@@ -119,7 +118,6 @@ import { TwoFactorAuthModule } from './modules/tfa/two-factor-auth.module';
     PlanModule,
     WorkspacesModule,
     GatewayModule,
-    TwoFactorAuthModule,
   ],
   controllers: [],
   providers: [],
